Revert "Add an 'isort' & 'black' task."
