+++ pygradle-plugin/src/main/groovy/com/linkedin/gradle/python/util/StandardTextValues.java
@@ -31,7 +31,6 @@ public enum StandardTextValues {
     CONFIGURATION_VENV("venv"),
     CONFIGURATION_WHEEL("wheel"),
     CONFIGURATION_FLAKE8("flake8"),
-    TASK_BLACK("runBlack"),
     TASK_BUILD_DOCS("buildDocs"),
     TASK_CLEAN_SAVE_VENV("cleanSaveVenv"),
     TASK_CHECK("check"),
@@ -43,7 +42,6 @@ public enum StandardTextValues {
     TASK_INSTALL_PROJECT("installProject"),
     TASK_INSTALL_PYTHON_REQS("installPythonRequirements"),
     TASK_INSTALL_TEST_REQS("installTestRequirements"),
-    TASK_ISORT("runIsort"),
     TASK_MYPY("runMypy"),
     TASK_PACKAGE_DOCS("packageDocs"),
     TASK_PACKAGE_JSON_DOCS("packageJsonDocs"),
